Finish Delete modal and prettify UI.
* put Modal as a variable in Menu to access openModal.
* style for light mode.

missing
* style system
* motion transition - nice to have.
* delete loading icon + optimistic delete.
