+++ src/components/ui/Menu.tsx
@@ -1,24 +1,90 @@
-import { Menu, Transition } from "@headlessui/react";
-import { forwardRef, Fragment, RefObject, useRef, type SVGProps } from "react";
+import { Menu, Transition, Dialog } from "@headlessui/react";
+import { Fragment, useState, type SVGProps } from "react";
 import { Bars3Icon } from "@heroicons/react/20/solid";
-import Modal from "./modal";
-import type { ModalRef } from "./modal";
-import { type Post } from "@prisma/client";
 
 export interface MenuProps extends React.HTMLAttributes<HTMLDivElement> {
-  id: Post["id"];
   onDelete: () => void;
 }
 
 const MenuDropDown: React.FC<MenuProps> = ({ ...props }) => {
-  const modal: RefObject<ModalRef> = useRef(null);
+  const [isOpen, setIsOpen] = useState(false);
+
+  function closeModal() {
+    setIsOpen(false);
+  }
+
+  function openModal() {
+    setIsOpen(true);
+  }
+
+  const modal = (
+    <Transition appear show={isOpen} as={Fragment}>
+      <Dialog as="div" className="relative z-10" onClose={closeModal}>
+        <Transition.Child
+          as={Fragment}
+          enter="ease-out duration-300"
+          enterFrom="opacity-0"
+          enterTo="opacity-100"
+          leave="ease-in duration-200"
+          leaveFrom="opacity-100"
+          leaveTo="opacity-0"
+        >
+          <div className="fixed inset-0 bg-black bg-opacity-60" />
+        </Transition.Child>
+
+        <div className="fixed inset-0 overflow-y-auto">
+          <div className="flex min-h-full items-center justify-center p-4 text-center">
+            <Transition.Child
+              as={Fragment}
+              enter="ease-out duration-300"
+              enterFrom="opacity-0 scale-95"
+              enterTo="opacity-100 scale-100"
+              leave="ease-in duration-200"
+              leaveFrom="opacity-100 scale-100"
+              leaveTo="opacity-0 scale-95"
+            >
+              <Dialog.Panel className="w-full max-w-md transform overflow-hidden rounded-2xl bg-white p-6 text-left align-middle shadow-xl transition-all">
+                <Dialog.Title
+                  as="h3"
+                  className="text-lg font-medium leading-6 text-gray-900"
+                >
+                  Are you sure you want to delete this Post?
+                </Dialog.Title>
+
+                <div className="mt-4 flex gap-4">
+                  <button
+                    onClick={() => {
+                      closeModal();
+                    }}
+                    className="inline-flex justify-center rounded-md border border-transparent bg-gray-500 px-4 py-2 text-sm font-medium text-gray-100 hover:bg-gray-400 focus:outline-none focus-visible:ring-2 focus-visible:ring-gray-800 focus-visible:ring-offset-2"
+                  >
+                    Cancel
+                  </button>
+                  <button
+                    type="button"
+                    className="inline-flex justify-center rounded-md border border-transparent bg-red-700 px-4 py-2 text-sm font-medium text-red-100 hover:bg-red-500 focus:outline-none focus-visible:ring-2 focus-visible:ring-red-500 focus-visible:ring-offset-2"
+                    onClick={() => {
+                      closeModal();
+                      props.onDelete();
+                    }}
+                  >
+                    Delete
+                  </button>
+                </div>
+              </Dialog.Panel>
+            </Transition.Child>
+          </div>
+        </div>
+      </Dialog>
+    </Transition>
+  );
 
   return (
     <Menu as="div" className="relative inline-block">
-      <Modal ref={modal}></Modal>
+      {modal}
       <Menu.Button className="inline-flex w-full justify-center rounded-md bg-black bg-opacity-0 px-1.5 py-1.5 text-sm font-medium text-white hover:bg-opacity-30 focus:outline-none focus-visible:ring-2 focus-visible:ring-white focus-visible:ring-opacity-75">
         <Bars3Icon
-          className=" h-4 w-4 text-violet-200 opacity-0 hover:text-violet-100 group-hover:opacity-100"
+          className=" h-4 w-4 text-gray-600 opacity-0 hover:text-violet-100 group-hover:opacity-100"
           aria-hidden="true"
         />
       </Menu.Button>
@@ -31,13 +97,13 @@ const MenuDropDown: React.FC<MenuProps> = ({ ...props }) => {
         leaveFrom="transform opacity-100 scale-100"
         leaveTo="transform opacity-0 scale-95"
       >
-        <Menu.Items className="absolute right-0 mt-2 w-56 origin-top-right divide-y divide-gray-100 rounded-md bg-white shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none">
+        <Menu.Items className="absolute right-0 mt-2 w-56 origin-top-right divide-y divide-gray-200 rounded-md bg-gray-300 shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none">
           <div className="px-1 py-1 ">
             <Menu.Item>
               {({ active }) => (
                 <button
                   className={`${
-                    active ? "bg-violet-500 text-white" : "text-gray-900"
+                    active ? "bg-gray-400 text-white" : "text-gray-900"
                   } group flex h-10 w-full items-center rounded-md px-2 py-2 text-sm`}
                 >
                   {active ? (
@@ -61,9 +127,9 @@ const MenuDropDown: React.FC<MenuProps> = ({ ...props }) => {
               {({ active }) => (
                 <button
                   className={`${
-                    active ? "bg-violet-500 text-white" : "text-gray-900"
+                    active ? "bg-red-400 text-white" : "text-gray-900"
                   } group flex h-10 w-full items-center rounded-md px-2 py-2 text-sm`}
-                  onClick={props.onDelete}
+                  onClick={() => openModal()}
                 >
                   <div
                     className="absolute flex"
@@ -71,12 +137,12 @@ const MenuDropDown: React.FC<MenuProps> = ({ ...props }) => {
                   >
                     {active ? (
                       <DeleteActiveIcon
-                        className="mr-2 h-5 w-5 text-violet-400"
+                        className="mr-2 h-5 w-5 text-pink-400"
                         aria-hidden="true"
                       />
                     ) : (
                       <DeleteInactiveIcon
-                        className="mr-2 h-5 w-5 text-violet-400"
+                        className="mr-2 h-5 w-5 text-pink-400"
                         aria-hidden="true"
                       />
                     )}
@@ -91,6 +157,7 @@ const MenuDropDown: React.FC<MenuProps> = ({ ...props }) => {
     </Menu>
   );
 };
+
 function EditInactiveIcon(props: SVGProps<SVGSVGElement>) {
   return (
     <svg
@@ -101,8 +168,8 @@ function EditInactiveIcon(props: SVGProps<SVGSVGElement>) {
     >
       <path
         d="M4 13V16H7L16 7L13 4L4 13Z"
-        fill="#EDE9FE"
-        stroke="#A78BFA"
+        fill="#FFFFFF"
+        stroke="#AAAAAA"
         strokeWidth="2"
       />
     </svg>
@@ -119,8 +186,8 @@ function EditActiveIcon(props: SVGProps<SVGSVGElement>) {
     >
       <path
         d="M4 13V16H7L16 7L13 4L4 13Z"
-        fill="#8B5CF6"
-        stroke="#C4B5FD"
+        fill="#AAAAAA"
+        stroke="#FFFFFF"
         strokeWidth="2"
       />
     </svg>
@@ -140,12 +207,12 @@ function DeleteInactiveIcon(props: SVGProps<SVGSVGElement>) {
         y="6"
         width="10"
         height="10"
-        fill="#EDE9FE"
-        stroke="#A78BFA"
+        fill="#FFFFFF"
+        stroke="#FF7777"
         strokeWidth="2"
       />
-      <path d="M3 6H17" stroke="#A78BFA" strokeWidth="2" />
-      <path d="M8 6V4H12V6" stroke="#A78BFA" strokeWidth="2" />
+      <path d="M3 6H17" stroke="#FF7777" strokeWidth="2" />
+      <path d="M8 6V4H12V6" stroke="#FF7777" strokeWidth="2" />
     </svg>
   );
 }
@@ -163,12 +230,12 @@ function DeleteActiveIcon(props: SVGProps<SVGSVGElement>) {
         y="6"
         width="10"
         height="10"
-        fill="#8B5CF6"
-        stroke="#C4B5FD"
+        fill="#FF7777"
+        stroke="#FFFFFF"
         strokeWidth="2"
       />
-      <path d="M3 6H17" stroke="#C4B5FD" strokeWidth="2" />
-      <path d="M8 6V4H12V6" stroke="#C4B5FD" strokeWidth="2" />
+      <path d="M3 6H17" stroke="#FFFFFF" strokeWidth="2" />
+      <path d="M8 6V4H12V6" stroke="#FFFFFF" strokeWidth="2" />
     </svg>
   );
 }
